feat(config): allow setting the get page timeout globally from the config
To change the timeout for how long a page is allowed to stall on `browser.get`, change
`getPageTimeout: timeout_in_millis` in the configuration. As before, you may also
change the timeout for one particular `get` call by using a second parameter:
`browser.get(url, timeout_in_sec)`

Breaking Change:

This change contains a small breaking change for consistency. Previously,
the second parameter to `get` changed the timeout in seconds. Now, the units
are milliseconds. This is consistent with all the other timeouts, as well as
base JavaScript functions like setTimeout.

 - before: `browser.get(url, 4)`
 - after: `browser.get(url, 4000)`