use the root element only to find the testability API, not scope searches
n 9a8f45a a change was introduced which made
Protractor's custom locators (by.binding, by.model, etc) use config.rootElement
as the root for all their searches. This meant that config.rootElement was used
both to specify how to get hold of Angular's injector as well as where to begin
searching for elements. This does not work for all cases, for example if a dialog
should be searched for elements but is a sibling, not a child, of ng-app.

This reverts that change, and uses document as the parent for all searches. This
is consistent with the behavior of the native locators by.id, by.css, and friends,
which do not scope their search based on config.rootElement.