feat(runner): add browser.getProcessedConfig method
Now, instances of the `browser` object have a `getProcessedConfig` method
which returns a promise that resolves to the current Protractor configuration
object for the current runner instance. This means that if multiCapabilities
are being used or tests are sharded, `getProcessedConfig` will return
an object with the `capabilities` and `specs` property specific to the
current instance.

Closes #1724