feat(typescript): adding typescript to protractor
Converting a 3 files over to typescript.

Adding an `npm prepublish` step that will use gulp to download the typings, transpile the files with tscto the built/ directory and copy the rest of the javascript files from lib/ to the built/ folder.

Also adding scripts to package.json for `npm run tsc` and `npm run tsc:w` for transpiling help.