chore(configParser): allow non-glob file pattern (#2754)
Cucumber allows line numbers to be passed in the filename in the form of `features/some.feature:42`. Glob expanding that results in an empty array and nothing being passed to the framework runner. This change checks for glob magic characters and only tries expanding it if found. Otherwise it just passes the filename verbatim. This was previously handled in [#2445] by stripping the line number first, but this is a more generic (non-cucumber) way to do it.

Glob needed to be upgraded for this which resulted in a weird [npm 3 bug] (https://github.com/npm/npm/issues/10637). Removing the rimraf package resolved this. It was only used to generate documentation which itself was removed a while ago.