chore(config): Update docs regarding proxies (#5048)
This change updates documentation on the different proxies.

The sauceAgent key's comment said that adding a webDriverProxy would override this and this is not true. sauceAgent is overwritten by sauceProxy since both are passed onto the saucelabs module.

This means that if you are under a corporate proxy and need to proxy both Sauce Labs API calls as well as calls to the Selenium endpoint (ondemand.saucelabs.com) both sauceProxy/sauceAgent and webDriverProxy need to be used.